requirements.txt
#Django<1.11.0
Django==1.11.29
django-extensions==1.9.8
#git+https://github.com/dbca-wa/ledger#egg=ledger
#git+https://github.com/dbca-wa/ledger@v2022.08.19.01#egg=ledger
git+https://github.com/dbca-wa/ledger#egg=ledger
# last stable commit for ledger tagged v2021.03.05.03 pr#836
#git+https://github.com/dbca-wa/ledger@4818666a1bda5fac274e9d0b48008797f04f6545
git+https://github.com/dbca-wa/dpaw-utils.git@0.3a16#egg=dpaw-utils
git+https://github.com/dbca-wa/dbca-utils@1.0.0
#psycopg2-binary==2.8.6
psycopg2==2.8.6
#git+https://github.com/django-oscar/django-oscar.git@17a1f6b17aba19e3c0b095e90acc654d463e0cf9#egg=django-oscar
git+https://github.com/django-oscar/django-oscar.git@8a3288da439cc2a878f44ae5c5101043e658d2a2#egg=django-oscar
social-auth-app-django==2.1.0
git+https://github.com/scottp-dpaw/social-core.git@email_fix#egg=social-auth-core
coverage==4.3.1
coveralls==1.1
#reportlab==3.5.26
reportlab==3.5.55
#django_bootstrap3==7.1.0
django_bootstrap3==12.1.0
django-braces>=1.8.1
django-datatables-view==1.13.0
django-reversion==3.0.0
django-reversion-compare==0.8.6
#git+https://github.com/bruth/django-preserialize.git@11f9e822250ea3374d5cfba477a00ca886689d9f#egg=django-preserialize
git+https://github.com/xzzy/django-preserialize.git#egg=django-preserialize
#django-countries==3.4.1
django-countries==6.1.3
django-cron==0.5.0
django-dynamic-fixture==1.9.1
openpyxl==2.4.8
datapackage==0.8.1
jsontableschema==0.10.1
python-dateutil==2.6.1
py4j==0.10.2.1
djangorestframework==3.8.2
djangorestframework-gis==0.14
djangorestframework-csv==2.0.0
pycountry==17.1.8
six>=1.10.0
django-ical==1.4
django-taggit==0.21.3
geojson==1.3.3
unicodecsv==0.14.1
PyPDF2==1.26.0
django-extra-fields==0.9
django-crispy-forms==1.7.0
django-reset-migrations==0.3.1
django-ckeditor==5.4.0
djangorestframework-datatables==0.4.0
django-confy==1.0.4
xlwt==1.3.0
xlsxwriter==1.1.2
django-concurrency==2.0
gunicorn==22.0.0
dj-static==0.0.6
mixer==6.1.3
django-dirtyfields==1.3.1
# django-smart-selects==1.5.4
# fix for https://github.com/jazzband/django-smart-selects/issues/253
git+https://github.com/digi604/django-smart-selects.git@js-unlinting-fixes
django-debug-toolbar==1.9
dj-database-url==0.5.0
werkzeug==0.14
BeautifulSoup4==4.9.1
phonenumbers==8.12.15
django-treebeard==4.3.1
#django-ckeditor
python-docx==0.8.10
docxtpl==0.11.3
django-multiselectfield==0.1.12
whitenoise==5.3.0
git+https://github.com/dbca-wa/django-media-serv.git#egg=django_media_serv

python-decouple==3.8
git+https://github.com/dbca-wa/appmonitor_client.git#egg=appmonitor_client

# For ER Diagram image generation 'python manage_wc.py graph_models --pygraphviz -a -g -I EmailUser,Application,Licence -o ~/Downloads/wlc_er_16Jun2023_v1.png'
pyparsing==2.4.6
pydot==1.4.2
pygraphviz==1.7
pydotplus==2.0.2

wheel==0.42.0
urllib3==2.0.7
requests==2.31.0

numpy==1.21.6
idna==3.7

setuptools==68.0.0
